Update conventions to follow existing cloud-functions
